Allow editing of bang triggers
Due to how the bang edit form worked, if you changed the trigger for a bang,
it would create a whole new bang while leaving the old one untouched. This was
useful for cloning bangs before I added the clone link for specifically that
purpose, but now that I have added the clone link, the old bang should be
deleted.