Add bang sharing
This is the last of the features I originally thought of adding to mallard,
so I guess in a sense this marks version 1.0. Custom bangs are nice and all,
but what's the point in having them if you can't share them? When a `share`
query parameter is passed alongside a search for "!settings", the `share`
query parameter (which is base64url-encoded JSON) is decoded and opened in
the bang editor. Clicking the "share" link on a bang in your settings menu
will create these links and copy them to your clipboard for easy sharing.